Multiple local search
Evidence seems to suggest that sometimes, the best topology given by the
grid search is not the best after running the local optimiser. In this
sense, the best topologies (these which gives S11 below -10dB) are
considered as valid candidates. This means that the local optimiser will
be run for each of them. This seems to improve the overall performance